refactor(frontend): reorganize imports and restore MetadataProgressUpdateForm component
