+++ apps/frontend/app/components/dashboard/forms/review-entity-form.tsx
@@ -33,7 +33,6 @@ import { useState } from "react";
 import { Form } from "react-router";
 import { match } from "ts-pattern";
 import { withQuery } from "ufo";
-import { convertThreePointSmileyToDecimal } from "~/components/dashboard";
 import {
 	ThreePointSmileyRating,
 	convertDecimalToThreePointSmiley,
@@ -46,6 +45,7 @@ import {
 	useUserPreferences,
 } from "~/lib/hooks";
 import { useReviewEntity } from "~/lib/state/media";
+import { convertThreePointSmileyToDecimal } from "../utils";
 
 export const ReviewEntityForm = ({
 	closeReviewEntityModal,
