--- apps/frontend/app/components/dashboard/index.tsx
@@ -1,11 +1,0 @@
-export * from "./error-boundary";
-export * from "./forms/add-entity-to-collections-form";
-export * from "./forms/create-measurement-form";
-export * from "./forms/review-entity-form";
-export * from "./modals/layout-modals";
-export * from "./forms/metadata-progress-update";
-export * from "./navigation/footer";
-export * from "./navigation/links-group";
-export * from "./navigation/navigation-config";
-export * from "./types";
-export * from "./utils";
