--- apps/frontend/app/components/fitness.action/index.tsx
@@ -1,12 +1,0 @@
-export * from "./exercise-display/display";
-export * from "./header";
-export * from "./hooks";
-export * from "./miscellaneous";
-export * from "./modals";
-export * from "./reorder";
-export * from "./rest-timer";
-export * from "./set-display/display";
-export * from "./stat-display-and-input";
-export * from "./supersets";
-export * from "./timer-and-stopwatch-drawer";
-export * from "./utils";
