--- apps/frontend/app/components/media-item/index.tsx
@@ -1,13 +1,0 @@
-export * from "./constants";
-export * from "./types";
-export * from "./utils";
-export * from "./displays/history-item";
-export * from "./displays/metadata-creator";
-export * from "./displays/podcast-episode";
-export * from "./displays/season-episode-details";
-export * from "./displays/show-episode";
-export * from "./displays/show-season";
-export * from "./displays/video-iframe";
-export * from "./modals/edit-history-modal";
-export * from "./modals/merge-metadata-modal";
-export * from "./modals/show-season-episodes-modal";
